all basic needs tests pass

+++ engine/actors/grow.cpp
@@ -101,7 +101,7 @@ Percent CanGrow::growthPercent() const
 void CanGrow::scheduleEvent(Area& area)
 {
 	ActorIndex actor = m_actor.getIndex();
-	Step delay = Step::create(std::round(AnimalSpecies::getStepsTillFullyGrown(area.getActors().getSpecies(actor)).get() / 100.f));
+	Step delay = Step::create(std::round((float)AnimalSpecies::getStepsTillFullyGrown(area.getActors().getSpecies(actor)).get() / 100.f));
 	m_event.resume(delay, area, *this);
 }
 void CanGrow::update(Area& area)
